requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.0.0
snowflake-connector-python>=3.0.0
